setting data in Recycler View remade general form of app

+++ app/src/main/java/com/example/weatherapp/data/network/OpenWeatherApiHelper.kt
@@ -14,8 +14,8 @@ interface OpenWeatherApiHelper {
     ): Observable<CurrentWeatherResponse>
 
     fun getForecastWeather(
-        lat: Int,
-        lon: Int,
+        lat: Double,
+        lon: Double,
         key:String,
         units:String,
         language:String
