+++ app/src/main/java/com/example/weatherapp/data/network/OpenWeatherApiService.kt
@@ -23,8 +23,8 @@ interface OpenWeatherApiService {
 
     @GET("forecast")
     fun getForecastWeather(
-        @Query("lat") lat:Int,
-        @Query("lon") lon:Int,
+        @Query("lat") lat:Double,
+        @Query("lon") lon:Double,
         @Query("appid") key:String,
         @Query("units") units:String,
         @Query("lang") language:String = "en"
